requirements.txt
affine==2.4.0
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
flower==2.0.1
fonttools==4.45.0
GDAL==3.4.1
geojson==3.1.0
geomet==1.1.0
gotrue==1.3.0
greenlet==3.0.1
h11==0.14.0
haversine==2.8.0
html2text==2020.1.16
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
humanize==4.9.0
idna==3.4
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
Pillow==10.1.0
postgrest==0.13.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rasterio==1.3.9
realtime==1.0.0
redis==5.0.1
requests==2.31.0
sentinelsat==1.2.1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
SQLAlchemy==2.0.23
starlette==0.27.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.1
tenacity==8.2.3
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
utm==0.7.0
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==10.4
wget==3.2
xmltodict==0.13.0
